% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_birdflow_mc.R
\name{calc_birdflow_mc}
\alias{calc_birdflow_mc}
\title{Function to calculate migratory connectivity (MC) on a BirdFlow object}
\usage{
calc_birdflow_mc(bf, ...)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{...}{
  Arguments passed on to \code{\link[BirdFlowR:lookup_timestep_sequence]{BirdFlowR::lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[BirdFlowR:lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[BirdFlowR:lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
The migratory connectivity of the BirdFlow model over the time period
indicated by \code{...}
}
\description{
This function calculates migratory connectivity based on the transitions
in a BirdFlowR model using the raster cells as regions.
It relies on \code{MigConnectivity::calcMC()} to do the bulk of the work.
}
\examples{
bf <- BirdFlowModels::amewoo
mc <- calc_birdflow_mc(bf, season = "prebreeding")
print(mc)

}
