% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_route_mc.R
\name{calc_route_mc}
\alias{calc_route_mc}
\title{Calc migratory connectivity from BirdFlowR routes}
\usage{
calc_route_mc(rts, bf)
}
\arguments{
\item{rts}{Output from \code{\link[BirdFlowR:route]{BirdFlowR::route()}}}

\item{bf}{The BirdFlow model used to make \code{rts}}
}
\value{
migratory connectivity estimated from the \code{rts} object.
}
\description{
\strong{WARNING} this function is experimental and new and likely to change
a bunch.
}
\details{
Currently, it assumes that all the routes start and end at the same
timestep and so uses the first and last timestep of the first route as the
starting and ending timesteps and then calculates the migratory connectivity
between those timesteps across all routes, dropping routes that don't
have both timesteps.

It treats the routes as "Telemetry" information when calling
\code{MigConnectivity} functions.

Current process is to call \code{MigConnectivity::estPsi} internally to calculate
the transition probabilities needed to calculate \code{MC}.
This step is really slow and then pass the result to \code{calcMC}.

It might make sense to instead call  \code{estMC} which I think would perform
both steps at once and output uncertainty.

Alternatively, we could see if we can call a more efficient function to
calculate the transitions probabilities as \code{estPsi} is doing a lot of
work to calculate uncertainties that aren't used with the current process.
}
