% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_birdflow_mc.R
\name{calc_birdflow_mc}
\alias{calc_birdflow_mc}
\title{Calculate migratory connectivity (MC) for a BirdFlow model.}
\usage{
calc_birdflow_mc(bf, from_marginals = TRUE, ...)
}
\arguments{
\item{bf}{A BirdFlow model}

\item{from_marginals}{Use TRUE (the default) to sample from distributions derived from
the fitted model parameters stored in the marginals. Use FALSE to use distributions
derived directly from eBird Status and Trends when sampling starting locations.}

\item{...}{
  Arguments passed on to \code{\link[BirdFlowR:lookup_timestep_sequence]{BirdFlowR::lookup_timestep_sequence}}
  \describe{
    \item{\code{season}}{a season name, season alias, or "all".  See
\code{\link[BirdFlowR:lookup_season_timesteps]{lookup_season_timesteps()}} for options.}
    \item{\code{start}}{The starting point in time specified as a
timestep, character date, or date object.}
    \item{\code{end}}{The ending point in time as a date or timestep.}
    \item{\code{direction}}{Either "forward" or "backward" defaults to
\code{"forward"} if not processing dates.  If using date input \code{direction} is
optional and is only used to verify the direction implicit in the dates.}
    \item{\code{season_buffer}}{Only used with \code{season} input. \code{season_buffer} is
passed to \code{\link[BirdFlowR:lookup_season_timesteps]{lookup_season_timesteps()}} and defaults to 1; it is the number
of timesteps to extend the season by at each end.}
    \item{\code{n_steps}}{Alternative to \code{end}  The end will be \code{n_steps} away from
\code{start} in \code{direction}; and the resulting sequence will have \code{n_step}
transitions and \code{n_steps + 1}  timesteps.}
  }}
}
\value{
The migratory connectivity of the BirdFlow model over the time period
indicated by \code{...}
}
\description{
This function calculates migratory connectivity based on the transitions
in a BirdFlowR model using the raster cells as regions.
It calculates the migratory connectivity metric MC as defined in Cohen 2018.
MC represents an abundance-weighted correlation that is calculated between the origin
locations and target locations, taking into account all grid transitions for the
specified time period.
}
\examples{
# load a BirdFlow model
bf <- BirdFlowModels::amewoo
# calculate migratory connectivity across the spring migration season:
mc <- calc_birdflow_mc(bf, season = "prebreeding")
print(mc)
# calculate mc from week 10 to week 20:
calc_birdflow_mc(bf, start=10, end=20)
}
\references{
Cohen EB, Hostetler JA, Hallworth MT, Rushing CS, Sillett TS, Marra PP.
Quantifying the strength of migratory connectivity.
Methods in Ecology and Evolution. 2018 Mar;9(3):513-24.
\doi{10.1111/2041-210X.12916}
}
\seealso{
\code{\link[=calc_route_mc]{calc_route_mc()}}
}
